Make vec overloads not templated
Change the `vec` overloads so they are no longer templated.  Instead
there is a separate overload for each vector length.  This allows the
application to pass a "swizzle" object as a parameter which expects
`vec` because the swizzle object is implicitly convertible to `vec`.

I also changed the formatting of the table of generic types to use
"[source]" blocks instead of "[code]" inline format.  This looks better
and it also allowed me to reformat some of the cells that have many
`vec` types.  Without this reformat, these cells were too big to fit on
one page, which caused an error with the PDF formatter.